Add files via upload
As part of my Web Development and Design Internship, I successfully completed the required tasks within the one-month time frame. The first task required extensive use of HTML5 and CSS, but I also added a bit of JavaScript to enhance the functionality and make the project more dynamic. The second task took more time than expected, but the end result was fruitful, and I was able to accomplish all of my tasks within the given time frame.